BREAKING CHANGE: changement d'emplacement pour les methodes d'inscription et de connection
