+++ APIMarcheEtDeviens/APIMarcheEtDeviens/DTO/ConnectionRandonneurDTO.cs
@@ -1,0 +1,9 @@
+﻿namespace APIMarcheEtDeviens.Services
+{
+	public class ConnectionRandonneurDTO
+	{
+		public string Mail { get; set; } = string.Empty;
+
+		public string MotDePasse { get; set; } = string.Empty;
+	}
+}
