+++ APIMarcheEtDeviens/APIMarcheEtDeviens/DTO/RandonneurDTO.cs
@@ -5,7 +5,6 @@ namespace APIMarcheEtDeviens.Services
 {
 	public class RandonneurDTO
 	{
-		public Guid RandonneurId { get; set; }
 		public string Nom { get; set; } = string.Empty;
 		public string Prenom { get; set; } = string.Empty;
